--- src/main/java/com/olga/ExtendedSearchPage.java
@@ -1,27 +1,0 @@
-package com.olga;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
-//import selenium.SearchPage;
-import com.olga.elements.Button;
-import com.olga.elements.TextField;
-
-public class ExtendedSearchPage /*implements SearchPage*/ {
-   /* @FindBy(id = "text")
-    private TextField searchField;
-
-    @FindBy(css = "input[type=\"submit\"]")
-    private Button searchButton;
-
-    @Override
-    public void search(final String query) {
-        searchField.clearAndType(query);
-        searchButton.click();
-    }
-
-    @Override
-    public void init(final WebDriver driver) {
-        PageFactory.initElements(new ExtendedFieldDecorator(driver), this);
-    }*/
-}
