--- src/main/java/com/olga/SearchForm.java
@@ -1,18 +1,0 @@
-package com.olga;
-
-import org.openqa.selenium.support.FindBy;
-import com.olga.elements.Button;
-import com.olga.elements.TextField;
-
-public class SearchForm extends AbstractContainer {
-    /*@FindBy(id = "text")
-    private TextField searchField;
-
-    @FindBy(css = "input[type=\"submit\"]")
-    private Button searchButton;
-
-    public void search(final String query) {
-        searchField.clearAndType(query);
-        searchButton.click();
-    }*/
-}
