--- src/main/java/com/olga/elements/Button.java
@@ -1,5 +1,0 @@
-package com.olga.elements;
-
-public interface Button extends Element {
-    void click();
-}
