--- src/main/java/com/olga/elements/TextField.java
@@ -1,11 +1,0 @@
-package com.olga.elements;
-
-public interface TextField extends Element {
-    void type(String text);
-
-    void clear();
-
-    void clearAndType(String text);
-
-    void waitForElement();
-}
