--- src/main/java/com/olga/elements/impl/AbstractElement.java
@@ -1,17 +1,0 @@
-package com.olga.elements.impl;
-
-import org.openqa.selenium.WebElement;
-import com.olga.elements.Element;
-
-abstract class AbstractElement implements Element {
-    protected final WebElement wrappedElement;
-
-    protected AbstractElement(final WebElement wrappedElement) {
-        this.wrappedElement = wrappedElement;
-    }
-
-    @Override
-    public boolean isDisplayed() {
-        return wrappedElement.isDisplayed();
-    }
-}
