+++ src/main/java/com/olga/elements/impl/Button.java
@@ -1,0 +1,18 @@
+package com.olga.elements.impl;
+
+import org.openqa.selenium.WebElement;
+
+public class Button extends CustomElement {
+    protected Button(final WebElement wrappedElement) {
+        super(wrappedElement);
+    }
+
+    public void click(){
+        wrappedElement.click();
+    }
+
+    @Override
+    public boolean isDisplayed() {
+        return false;
+    }
+}
